UNT-T23887: Refactor ARView
- Move components to separate files
- Add code documentations
