+++ app/src/main/java/com/simform/ssfurnicraftar/utils/constant/Constants.kt
@@ -31,10 +31,18 @@ object Constants {
     const val MODEL_BOUNCING_HEIGHT = 0.05F // Meter
     const val MODEL_BOUNCING_DURATION = 1000 // MS
     const val MODEL_NO_ROTATION = 0F // Degree
+    const val MODEL_360_ROTATION = 360F // Degree
+    const val MODEL_ROTATION_DURATION = 10000 // MS
     const val MODEL_NO_SCALE = 0F
     const val MODEL_NO_HEIGHT = 0F // Meter
     const val MODEL_RENDER_LOWEST_PRIORITY = 7
     const val MODEL_BASE_COLOR_INDEX = 0
+
+    /**
+     * Gesture
+     */
+    const val DEFAULT_GESTURE_OVERLAY_ITERATIONS = 1
+    const val PLAY_COUNT_ZERO = 0
 
     /**
      * Provider
@@ -49,6 +57,13 @@ object Constants {
     const val IMAGE_CAPTURE_HANDLER_NAME = "PixelCopier"
     const val IMAGE_FILE_EXTENSION = "jpeg"
 
-    // Lottie
+    /**
+     * Animation
+     */
     const val ANIM_COMPLETE_VALUE = 1F
+
+    /**
+     * Color Picker
+     */
+    const val ANIM_LABEL_COLOR_PICKER = "ColorPickerAnimation"
 }
